requirements.txt
asyncio==3.4.3
pyserial==3.5
websockets==15.0.1
